1ms计算x y位置 中断方式计算码盘脉冲 每次碰到下限位 清零脉冲计数

+++ USER/encoder.h
@@ -22,7 +22,7 @@ enum EncoderNUM{
 #define ENCODER_NUMS       		(ENCODER_ID_MAX - ENCODER_ID_MIN + 1)
 
 typedef struct _Encoder  {
-	s16 pluse;
+	s32 pluse;
 	//u16 diff;
 }Encoder_t;
